Added minior fixes to the implementation
