+++ components/cli/pkg/commands/login.go
@@ -19,38 +19,14 @@
 package commands
 
 import (
-	"context"
-	"crypto/tls"
-	"encoding/json"
 	"fmt"
-	"github.com/cellery-io/sdk/components/cli/pkg/constants"
+	"strings"
+
+	"github.com/nokia/docker-registry-client/registry"
+
 	"github.com/cellery-io/sdk/components/cli/pkg/registry/credentials"
 	"github.com/cellery-io/sdk/components/cli/pkg/util"
-	"github.com/jwt-go"
-	"github.com/nokia/docker-registry-client/registry"
-	"golang.org/x/crypto/ssh/terminal"
-	"io/ioutil"
-	"net"
-	"net/http"
-	"os"
-	"path/filepath"
-	"strconv"
-	"strings"
-	"time"
 )
-
-var loginConf Conf
-
-type Conf struct {
-	CallBackDefaultPort int    `json:"call-back-default-port"`
-	CallBackContextPath string `json:"call-back-context-path"`
-	CallBackHost        string `json:"call-back-host"`
-	CallBackParameter   string `json:"call-back-parameter"`
-	RedirectSuccessUrl  string `json:"redirect-success-url"`
-	IsHost              string `json:"is-host"`
-	IsPort              int    `json:"is-port"`
-	SpClientId          string `json:"sp-client-id"`
-}
 
 // RunLogin requests the user for credentials and logs into a Cellery Registry
 func RunLogin(registryURL string, username string, password string) {
@@ -88,7 +64,11 @@ func RunLogin(registryURL string, username string, password string) {
 		fmt.Println("Logging in with existing Credentials")
 	} else {
 		if password == "" {
-			registryCredentials.Username, registryCredentials.Password, err = requestCredentials(username)
+			if username == "" {
+				registryCredentials.Username, registryCredentials.Password, err = credentials.FromBrowser(username)
+			} else {
+				registryCredentials.Username, registryCredentials.Password, err = credentials.FromTerminal(username)
+			}
 			if err != nil {
 				util.ExitWithErrorMessage("Error occurred while reading Credentials", err)
 			}
@@ -124,159 +104,4 @@ func RunLogin(registryURL string, username string, password string) {
 
 	spinner.Stop(true)
 	util.PrintSuccessMessage(fmt.Sprintf("Successfully logged into Registry: %s", util.Bold(registryURL)))
-}
-
-// This function is to request credentials from the user
-func requestCredentials(username string) (string, string, error) {
-	initConf()
-	timeout := make(chan bool)
-	ch := make(chan string)
-	var code string
-	httpPortString := ":" + strconv.Itoa(loginConf.CallBackDefaultPort)
-	var codeReceiverPort = loginConf.CallBackDefaultPort
-	// This is to start the CLI auth in a different port is the default port is already occupied
-	for {
-		_, err := net.Dial("tcp", httpPortString)
-		if err != nil {
-			break
-		}
-		codeReceiverPort++
-		httpPortString = ":" + strconv.Itoa(codeReceiverPort)
-	}
-	var gitHubAuthUrl = "https://" + loginConf.IsHost + ":" + strconv.Itoa(loginConf.IsPort) +
-		"/oauth2/authorize?scope=openid&response_type=" + "code&redirect_uri=http%3A%2F%2F" +
-		loginConf.CallBackHost + "%3A" + strconv.Itoa(codeReceiverPort) +
-		"%2Fauth&client_id=" + loginConf.SpClientId + "&fidp=google"
-	fmt.Println("\n", gitHubAuthUrl)
-	go func() {
-		mux := http.NewServeMux()
-		server := http.Server{Addr: httpPortString, Handler: mux}
-		//var timer *time.Timer
-		mux.HandleFunc(loginConf.CallBackContextPath, func(w http.ResponseWriter, r *http.Request) {
-			err := r.ParseForm()
-			if err != nil {
-				util.ExitWithErrorMessage("Error parsing the code", err)
-			}
-			code = r.Form.Get(loginConf.CallBackParameter)
-			ch <- code
-			if err != nil {
-				fmt.Printf("Error occurred while establishing the server: %s\n", err)
-				return
-			}
-			if len(code) != 0 {
-				http.Redirect(w, r, "https://"+loginConf.IsHost+":"+strconv.Itoa(loginConf.IsPort)+
-					"/authenticationendpoint/auth_success.html", http.StatusSeeOther)
-			} else {
-				http.Redirect(w, r, "fail", http.StatusSeeOther)
-			}
-			flusher, ok := w.(http.Flusher)
-			if ok {
-				flusher.Flush()
-			}
-			err = server.Shutdown(context.Background())
-			if err != nil {
-				util.ExitWithErrorMessage("Error while shutting down the server\n", err)
-			}
-		})
-		if err := server.ListenAndServe(); err != nil && err != http.ErrServerClosed {
-			util.ExitWithErrorMessage("Error while establishing the service", err)
-		}
-	}()
-
-	err := util.OpenBrowser(gitHubAuthUrl)
-	if err != nil {
-		fmt.Printf("Could not resolve the given url %s. Started to operate in the healess "+
-			"mode\n", gitHubAuthUrl)
-		return handleHeadless(username)
-	}
-	// Setting up a timeout
-	go func() {
-		time.Sleep(5 * time.Minute)
-		timeout <- true
-	}()
-	// Wait for a code, or timeout
-	select {
-	case <-ch:
-	case <-timeout:
-		fmt.Println("Time out. Did not receive any code")
-		os.Exit(0)
-	}
-	token := getTokenFromCode(code, codeReceiverPort)
-	username, idToken := getSubjectAndUsernameFromJWT(token)
-	return username, idToken, nil
-}
-
-// This function is to allow this login flow to work in headless mode
-func handleHeadless(username string) (string, string, error) {
-	var password string
-	if username == "" {
-		var err error
-		fmt.Print("Enter username: ")
-		_, err = fmt.Scanln(&username)
-		if err != nil {
-			util.ExitWithErrorMessage("Error reading the input username", err)
-		}
-	}
-	fmt.Print("Enter password/token: ")
-	bytePassword, err := terminal.ReadPassword(0)
-	if err != nil {
-		util.ExitWithErrorMessage("Error reading the input token", err)
-	}
-	password = strings.TrimSpace(string(bytePassword))
-	fmt.Println()
-	return username, password, nil
-}
-
-// Returns the extracted subject from the JWT
-func getSubjectAndUsernameFromJWT(token string) (string, string) {
-	var result map[string]interface{}
-	err := json.Unmarshal([]byte(token), &result)
-	if err != nil {
-		util.ExitWithErrorMessage("Error while unmarshal the id_token", err)
-	}
-	idToken := result["id_token"]
-	jwtToken, _ := jwt.Parse(idToken.(string), nil)
-	claims := jwtToken.Claims.(jwt.MapClaims)
-	return claims["sub"].(string), idToken.(string)
-}
-
-// Returns the JWT from the auth code provided
-func getTokenFromCode(code string, port int) string {
-	tokenUrl := "https://" + loginConf.IsHost + ":" + strconv.Itoa(loginConf.IsPort) + "/oauth2/token"
-	responseBody := "client_id=" + loginConf.SpClientId + "&grant_type=authorization_code&code=" + code +
-		"&redirect_uri=http://localhost:" + strconv.Itoa(port) + loginConf.CallBackContextPath
-	body := strings.NewReader(responseBody)
-	// Token request
-	req, err := http.NewRequest("POST", tokenUrl, body)
-	if err != nil {
-		util.ExitWithErrorMessage("Error while creating the code receiving request", err)
-	}
-	req.Header.Add("Content-Type", "application/x-www-form-urlencoded;charset=UTF-8")
-	http.DefaultTransport.(*http.Transport).TLSClientConfig = &tls.Config{InsecureSkipVerify: true}
-	res, err := http.DefaultClient.Do(req)
-	if err != nil {
-		fmt.Printf("Could not connect to the client at %s", tokenUrl)
-		util.ExitWithErrorMessage("Error occurred while connecting to the client", err)
-	}
-	defer res.Body.Close()
-	respBody, err := ioutil.ReadAll(res.Body)
-	if err != nil {
-		util.ExitWithErrorMessage("Error occurred while reading the response body", err)
-	}
-	return string(respBody)
-}
-
-// Read the config.json file from cellery home
-func initConf() {
-	loginConf = Conf{}
-	configFilePath := filepath.Join(util.UserHomeDir(), constants.CELLERY_HOME, constants.CONFIG_FILE)
-	configFile, err := ioutil.ReadFile(configFilePath)
-	if err != nil {
-		fmt.Printf("Could not read from the file %s \n", configFilePath)
-		util.ExitWithErrorMessage("Error reading the file", err)
-	}
-	err = json.Unmarshal(configFile, &loginConf)
-	if err != nil {
-		util.ExitWithErrorMessage("Error while unmarshal the yaml", err)
-	}
 }
