+++ components/cli/pkg/constants/const.go
@@ -148,5 +148,3 @@ const CELLERY_SQL_PASSWORD = "cellery-sql-user"
 
 const INGRESS_MODE_NODE_PORT = "Node port [kubeadm, minikube]"
 const INGRESS_MODE_LOAD_BALANCER = "Load balancer [gcp, docker for desktop]"
-
-const CONFIG_FILE = "config.json"
